requirements.txt
fastapi
requests
python-dotenv
uvicorn
jinja2
python-multipart
sqlalchemy
os
